==> datawork/tweet_geoparse_algorithm/code/02_implement_algorithm.R <==
# Geoparse Tweets
# Implement Algorithm

# Implement algorithm, looping over different gazetteers in order to generate
# results using raw, augmented, augmented OSM, augmented Google Places and 
# augmented Geonames. For each gazetter, exports dataframe with tweets and 
# coordinates as generated by the algorithm

# Load Data --------------------------------------------------------------------
## Tweets
tweets_df <- readRDS(file.path(tweets_geoparse_dir, "processed_data", "tweets_for_geolocation.Rds"))

## Roads and Estates
roads <- readRDS(file.path(osm_dir, "data", "processed_data", "roads", "osm_roads_aug.Rds"))
estates <- readRDS(file.path(estates_dir, "data", "nairobi_estates.Rds"))

# Parameters -----------------------------------------------------------------
prepositions <- list(c("EVENT_WORD after", "EVENT_WORD near", "EVENT_WORD outside", 
                       "EVENT_WORD past", "around", "hapo", "just after", "just before", 
                       "just past", "near", "next to", "opposite", "outside", "past", 
                       "you approach", "apa", "apo", "hapa", "right after", 
                       "right before", "right past", "just before you reach"), 
                     c("EVENT_WORD at", "before"), 
                     c("after"),
                     c("at",
                       "happened at", "at the", "pale"), 
                     c("between", "from",
                       "btw", "btwn"), 
                     c("along", "approach", "in", "on", "opp", "to", "towards",
                       "toward") 
)

event <- c("accidents", "accident", "crash", "crush", "overturn", "overturned", 
           "collision", "wreck", "wreckage", "pile up", "pileup", "incident", 
           "hit and run", "hit", "roll", "rolled", "read end", "rear ended")
junction <- c("intersection", "junction")
false_positive <- c("githurai bus", "githurai matatu", 
                    "githurai 45 bus", "githurai 45 matatu",
                    "city hoppa bus", "hoppa bus",
                    "rongai bus", "rongai matatu", "rongai matatus",
                    "machakos bus", "machakos minibus", "machakos matatu",
                    "at ntsa kenya", 
                    "service lane", "star bus",
                    "prius", "mpya bus",
                    "heading towards") 
type_list <- list(c("bus_station","transit_station","stage_added", "stage", "bus_stop"),
                  c("mall", "shopping_mall"),
                  c("restaurant", "bakery", "cafe"),
                  c("building"),
                  c("parking"))

# Locate Crashes ---------------------------------------------------------------
# c("aug", "aug_geonames", "aug_google", "aug_osm", "raw")
for(gaz_type in c("aug_google", "aug_geonames", "aug_osm", "raw")){
  
  landmarks <- readRDS(file.path(landmarkgaz_dir, "data", "gazetteers",
                                 paste0("landmark_gazetter_",gaz_type,".Rds")))
  
  alg_out_sf <- locate_event(text = tweets_df$tweet,
                             landmark_gazetteer = landmarks, 
                             roads = roads, 
                             areas = estates, 
                             prepositions_list = prepositions, 
                             prep_check_order = "prep_then_pattern", 
                             event_words = event, 
                             junction_words = junction, 
                             false_positive_phrases = false_positive, 
                             type_list = type_list, 
                             clost_dist_thresh = 500,
                             fuzzy_match = TRUE,
                             fuzzy_match.min_word_length = c(5,11),
                             fuzzy_match.dist = c(1,2),
                             fuzzy_match.ngram_max = 3,
                             fuzzy_match.first_letters_same = TRUE,
                             fuzzy_match.last_letters_same = TRUE,
                             crs_distance = "+init=epsg:21037", 
                             crs_out = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
                             quiet = F,
                             mc_cores = 1)
  
  # Prep Outputs -----------------------------------------------------------------
  # Restrict to points and add to tweets
  alg_out_sf$status_id_str <- tweets_df$status_id_str
  alg_out_sf$geometry_type <- alg_out_sf$geometry %>% st_geometry_type() %>% as.character()
  alg_out_sf$geometry_dim <- alg_out_sf$geometry %>% st_dimension()
  
  alg_out_df <- alg_out_sf %>%
    filter(geometry_type %in% "POINT") %>%
    filter(!is.na(geometry_dim)) %>%
    dplyr::select(-text) %>%
    as("Spatial") %>%
    as.data.frame() %>%
    dplyr::rename(lon_alg = coords.x1,
                  lat_alg = coords.x2)
  
  tweets_alg_df <- merge(tweets_df, alg_out_df, by="status_id_str", all.x=T)
  
  # Export -----------------------------------------------------------------------
  saveRDS(tweets_alg_df, file.path(tweets_geoparse_dir, "processed_data", paste0("tweet_geoparse_gaz_",gaz_type,".Rds")))
}
